t12/toe/pyproject.toml
[tool.poetry]
name = "Tools For Engineers"
version = "0.1.0"
description = "Practicals for Tools For Engineers"
authors = ["ksdfg <ksdfg123@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
selenium = "^3.141.0"
pytest = "^6.2.2"
python-decouple = "^3.4"

[tool.poetry.dev-dependencies]
black = "^20.8b1"

[tool.black]
line-length = 120

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
